+++ internal/service/process_windows.go
@@ -1,0 +1,12 @@
+//go:build windows
+
+package service
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func configureDetached(command *exec.Cmd) {
+	command.SysProcAttr = &syscall.SysProcAttr{CreationFlags: 0x00000008 | 0x00000200}
+}